Change all JSDoc in ffl.js to use absolute imports, making library globals work
Necessary for UMD but also may undo this because it's just, ugly.
Thankfully, this change is actually far easier to undo than to bring it
on which is why I'm adding it.
... or maybe it will stay? If this approach is fully compatible with
TypeScript it may perhaps be worth using, don't know.

_applyModulateParam is the only function that got actual changes outside
of this in the update, just a note if this does get reverted.