Expand enums, fix hair flip with pAdjustMatrix, add TODO pathnames, misc JSDoc
Specifically FFLExpression and FFLResult enums.
Added a condition to set negative X scale for hair if the adjust matrix
seems to have that.
Annotate the file with TODO pathnames to use when the library is split
up into multiple files later on.
Add texture map to FFL/LUTShaderMaterialParameters types and move them
to the top so they are accessible outside of the IIFE.
Misc changes to documentation; convert some inline comments to JSDoc and
document FFLCharModelDesc, FFLModelFlag
Simplfify updateCharModelInScene in demo-minimal.js to not use
updateCharModel as it is not necessary.