Fix faceline order, blank expr, LUT fix, matClass check, handle no content-length
* Fix draw order of faceline texture which was always wrong.
* Ignore setting expression if blank expression is set (no mask mesh),
so that it can be set as a single expression (can't be set as one of
multiple expressions though like cat/dog)
* Fix typos in LUTShaderMaterial to use proper normalized normal,
leading to unexpected lighting especially when using flipped hair.
* Handle case where Content-Length does not exist on fetch response
(surprisingly common on web servers?) by falling back to reading as
ArrayBuffer, rather than throwing an error. Most of the time the
response is chunked, maybe that can be handled separately.