Allow using built-in materials if using a material without modulateMode.
Specifically...
* convertModelTexturesToRGBA will convert textures in a CharModel to
RGBA using a material that is compatible with modulateMode.
* If you call createCharModel with an unsupported material, and then
call initCharModelTextures with a material class that supports
modulateMode, the textures will be converted.
You're meant to use TextureShaderMaterial (or FFL/LUTShaderMaterials) if
you don't have the others.
That will help out with either WebGPU support or linear color space
support later down the line as well, assuming the FFL/LUT shaders
wouldn't be ported and the default Three.js shaders will have to be
used.

* There is also this new method that will convert all textures in the
CharModel to DataTexture (rather than being render targets):
convModelTargetsToDataTex

* Also fix FATAL OVERSIGHTS preventing non half float resources from
working on last commit and icons not updating in demo-basic oooooooops