Add ResourceLoadHelper.js for loading resource from file and error handling in demos.
ResourceLoadHelper.js is a JS module meant for the demos that will
inject an inline HTML UI that you can provide an FFL resource file to,
it'll also open them from zips (importing fflate). If more demos come, I
intend for them to use it.

That will actually let the demos load over file://, without needing a
web server - that is, except for the fact that it still needs to fetch()
ffl-emscripten.wasm. If you build with -s SINGLE_FILE=1 it will actually
fully function but that is not in this commit right now nor am
I sure I want to add it

This also removes initializeFFLWithResource since I felt it was
unnecessary, if you want the same functionality you should be able to
do: initializeFFL(fetch('hxxp://url'), module)