Add SampleShaderMaterial from Nintendo Switch
This is the elusive SampleShader.bnsh port to Three.js, from the
FFL-Testing shader. This is the most unique out of the three, needing
the most hacky workarounds.

1. It needs predefined material tables to use for each shape's color
index, as well as the indices to be available to set uniforms from.
- These get passed via the "colorInfo" object, created by the
CharModel, passed by ffl.js and demo-basic.js.
2. It needs the faceline color to be transparent.
  - Achieved by a static property on the material.
- This can be worked around, but beards will not appear correctly.
(May possibly be fixable)
3. In order for the glasses to match the way they look on the real
console, the normals need to be modified from their originals in
FFL/AFL's resources.
  - This needed the hackiest workaround.
- The material has a property that will modify the glass mesh's
BufferGeometry to add a new normal attribute.
- I'm not sure how to work around this, since normals are critical for
lighting.

The thing shared in common with all three of these is that the material
should not have to "know about" or interfere with these things. Any one
of these three  requirements breaks the isolation/encapsulation that
materials usually have.

With heavy modification, it may be possible to modify the shader to
allow 2 and 3 to be achieved with just modifying the shader code, but
reversing how the sssColor/specularColors were created algebraically or
otherwise will be pretty tough, especially if they are picked
artistically.
By far the biggest limitation is only being able to use the colors used
on Switch due to the material tables. It may be possible to use e.g.
common colors for potentially other elements, but that is to be seen.