Comment all console.debug lines, remove divider comments
* console.debug: These were useful for development, but realistically
speaking a "library" should not emit these.
- They were critical for addressing memory leaks, but I'm not aware of
a problem with that right now.
- The TextureManager logging was even more important. I chose to make
that optional since that REALLY got out of hand, but most JS dead code
elimination was too stupid to remove that.
- Perhaps I should add logging back as optional if I have to debug
this again, but with JS it's not as painless as having an #ifdef.

* The dividers are annoying to maintain and get out of date. I couldn't
make them anymore after I moved from Kate which I had a script for.